+++ packages/dart_edge_s3_client/lib/dart_edge_s3_client.dart
@@ -7,6 +7,7 @@ library;
 export 'src/dart_edge_s3_client.dart';
 export 'src/s3_client_config.dart';
 export 'src/s3_delete_object_result.dart';
+export 'src/s3_download_stream_counters.dart';
 export 'src/s3_get_object_bytes_result.dart';
 export 'src/s3_get_object_file_result.dart';
 export 'src/s3_get_object_stream_result.dart';